Extracts SyncDispatcher to a class
