+++ purchases/src/test/java/com/revenuecat/purchases/SyncDispatcher.kt
@@ -1,0 +1,24 @@
+package com.revenuecat.purchases
+
+import io.mockk.mockk
+import java.util.concurrent.RejectedExecutionException
+
+internal class SyncDispatcher : Dispatcher(mockk()) {
+
+    private var closed = false
+
+    override fun enqueue(call: AsyncCall) {
+        if (closed) {
+            throw RejectedExecutionException()
+        }
+        call.run()
+    }
+
+    override fun close() {
+        closed = true
+    }
+
+    override fun isClosed(): Boolean {
+        return closed
+    }
+}
